migrate EvriDeliveryService, move Status enum names to the enum itself
