+++ app/src/main/java/dev/itsvic/parceltracker/ui/views/ParcelView.kt
@@ -39,7 +39,6 @@ import dev.itsvic.parceltracker.api.ParcelHistoryItem
 import dev.itsvic.parceltracker.api.Service
 import dev.itsvic.parceltracker.api.Status
 import dev.itsvic.parceltracker.api.getDeliveryServiceName
-import dev.itsvic.parceltracker.api.statusToHumanString
 import dev.itsvic.parceltracker.ui.components.ParcelHistoryItemRow
 import dev.itsvic.parceltracker.ui.theme.ParcelTrackerTheme
 import java.time.LocalDateTime
@@ -128,13 +127,12 @@ fun ParcelView(
                 }
             }
 
-            statusToHumanString[parcel.currentStatus]?.let {
-                item {
-                    Text(
-                        LocalContext.current.getString(it),
-                        style = MaterialTheme.typography.headlineLarge
-                    )
-                }
+
+            item {
+                Text(
+                    LocalContext.current.getString(parcel.currentStatus.nameResource),
+                    style = MaterialTheme.typography.headlineLarge
+                )
             }
 
             items(parcel.history) { item ->
